% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_orcid.R
\name{add_orcid}
\alias{add_orcid}
\title{add_orcid <- function(input_file) {}
\usage{
add_orcid(input_file)
}
\arguments{
\item{input_file}{file including full path of all Bfile for which
DM operator information will be added in the file}
}
\value{
Bfile including DM operator information
}
\description{
This function add information regarding the dm operator
}
