% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOXY_adj.R
\name{DOXY_adj}
\alias{DOXY_adj}
\title{DOXY_adj}
\usage{
DOXY_adj(
  filenc_core,
  filenc,
  PARAM_name,
  OFFSET,
  SLOPE,
  DRIFT,
  INCLINE_T,
  ERROR,
  profile_date_juld,
  launch_date_juld
)
}
\arguments{
\item{filenc_core}{An object of class ncdf4 of argo core file}

\item{filenc}{An object of class ncdf4 of argo bfile}

\item{PARAM_name}{name of the variable to be adjusted (string)}

\item{OFFSET}{offset value for the oxygen adjustment (numeric)}

\item{SLOPE}{slope value for the oxygen adjustment (numeric)}

\item{DRIFT}{drift value for the oxygen adjustment (numeric)}

\item{INCLINE_T}{incline T value for the oxygen adjustment (numeric)}

\item{ERROR}{vector oxygen error values (numeric)}

\item{profile_date_juld}{julian day value (numeric) for the profile}

\item{launch_date_juld}{julian day value of the float launch date}
}
\value{
vector adjusted oxygen values (numeric)
}
\description{
This function estimates oxygen adjustment according to
https://doi.org/10.13155/46542
}
