% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NITRATE_ERROR_ESTIMATION.R
\name{NITRATE_ERROR_ESTIMATION}
\alias{NITRATE_ERROR_ESTIMATION}
\title{DOXY_adj}
\usage{
NITRATE_ERROR_ESTIMATION(filenc, NITRATE, NITRATE_ADJUSTED, ERROR, index_param)
}
\arguments{
\item{filenc}{An object of class ncdf4 of argo bfile}

\item{NITRATE}{vector of nitrate raw values (numeric)}

\item{NITRATE_ADJUSTED}{vector of nitrate adjusted values (numeric)}

\item{ERROR}{nitrate error value}

\item{index_param}{parameter index in the file matrix}
}
\value{
vector of adjusted nitrate error value
}
\description{
This function estimates nitrate error according to
https://doi.org/10.13155/84370
}
