% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROFILE_PARAM_QC_ESTIMATION.R
\name{profile_param_qc_estimation}
\alias{profile_param_qc_estimation}
\title{profile_param_qc_estimation}
\usage{
profile_param_qc_estimation(
  filenc,
  param_qc,
  profile_param_qc_name,
  profile_param_qc,
  n_qc,
  i_prof_param
)
}
\arguments{
\item{filenc}{An object of class ncdf4 of argo file}

\item{param_qc}{matrix of parameter qc to be changes}

\item{profile_param_qc_name}{name of parameter qc to be estimated}

\item{profile_param_qc}{matrix of parameter qc to be estimated}

\item{n_qc}{number of QC for param}

\item{i_prof_param}{profile index of the parameter in the nc file}
}
\value{
update history section of ncfile
}
\description{
calculate the profile_param_qc after a change
}
