% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REPORT_QC.R
\name{REPORT_QC}
\alias{REPORT_QC}
\title{REPORT_QC}
\usage{
REPORT_QC(input_file)
}
\arguments{
\item{input_file}{text file including all usefull information
for the adjustment separated by ";"
filename             full path of the B file
var_level            name of the level variable
level_value          level value
param                name of the variable for wich _QC should be changed
old_value            old QC param value
new_value            new QC param value
corr_value           corresponding param value
date_update          date update yyyymmddhhmmss
mandating_institution institution that mandates que DMQC}
}
\value{
Bfile with param_qc and profile_param_qc changed
}
\description{
This program is designed to change RAW QC
}
