% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HISTORY_UPDATE.R
\name{history_update}
\alias{history_update}
\title{history_update}
\usage{
history_update(
  filenc,
  param,
  i_prof_param,
  step,
  date_update,
  institution = NULL
)
}
\arguments{
\item{filenc}{An object of class ncdf4 of argo file}

\item{param}{argo parameter name (string)}

\item{i_prof_param}{profile index of the parameter in the nc file}

\item{step}{"DM" delayed mode action or "QC" Quality change action}

\item{date_update}{update text with format yyyymmddHHMMSS}

\item{institution}{institution code following NERC R04 table}
}
\value{
update history section of ncfile
}
\description{
update the history section in the netcdf file
}
